+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 ## [0.1.1] - 2023-07-12
 
 - Threads with Image
+- Threads with Link Attachment
 - Fixed major issues